Legg til dato og flytt komponenter til egen fil
